Move fp16 matrix mult unittest to compute tests
